# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "bile-mo API",
        "description": "bile-mo is a web service offering a selection of mobile phones",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/bile-mo-api/v1",
            "description": "bile-mo API main server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products management"
                ],
                "summary": "Returns list of products",
                "operationId": "2d4d6e47a3073543b52806d69bc60570",
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Ok",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products management"
                ],
                "summary": "Create a new product",
                "operationId": "52743ff1d0efd483c592034857ec62c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {},
                                    "description": {},
                                    "color": {},
                                    "country": {
                                        "type": "object"
                                    },
                                    "brand": {
                                        "type": "object"
                                    },
                                    "memory": {
                                        "type": "object"
                                    },
                                    "user": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Created",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "HTTP_CONFLICT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Conflict",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Bad request",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products management"
                ],
                "summary": "Returns product by id",
                "operationId": "386c548c4b9588fefece38298652ce2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Ok",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products management"
                ],
                "summary": "Updates a product by id",
                "operationId": "683218ba4e79dd7932ca7747dc4b94c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {},
                                    "description": {},
                                    "color": {},
                                    "country": {
                                        "type": "object"
                                    },
                                    "brand": {
                                        "type": "object"
                                    },
                                    "memory": {
                                        "type": "object"
                                    },
                                    "user": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Created",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductDTO"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "HTTP_CONFLICT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Conflict",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Bad request",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products management"
                ],
                "summary": "Deletes a product by id",
                "operationId": "8edd2225424c38a80c648d4be60f13b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "HTTP_NO_CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "No content",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BrandDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "ColorDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "CountryDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "MemoryDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "memoryCapacity": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": false
                    },
                    "description": {
                        "type": "text",
                        "nullable": false
                    },
                    "country": {
                        "$ref": "#/components/schemas/CountryDTO"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/BrandDTO"
                    },
                    "memory": {
                        "$ref": "#/components/schemas/MemoryDTO"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDTO"
                    },
                    "colors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ColorDTO"
                        },
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "UserDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string",
                        "nullable": false
                    },
                    "password": {
                        "type": "string",
                        "nullable": false
                    },
                    "customer": {
                        "type": "array"
                    },
                    "roles": {
                        "type": "array",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        }
    }
}